requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
google-api-python-client==2.170.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
tenacity==9.0.0
structlog==24.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
